+++ src/main/java/be/ugent/knows/idlabFunctions/state/SimpleInMemoryMapState.java
@@ -16,6 +16,24 @@ public class SimpleInMemoryMapState implements MapState {
     private final static Logger log = LoggerFactory.getLogger(SimpleInMemoryMapState.class);
     private final static Map<String, Map<String, String>> stateFileToMap = new HashMap<>();
 
+    /**
+     * Loads or creates a new state map for the given state file path.
+     * Associates the specified value with the specified key in this map
+     * (optional operation).  If the map previously contained a mapping for
+     * the key, the old value is replaced by the specified value.
+     * <br>
+     * Note that the state is only persisted to file upon {@link #close()}.
+     *
+     * @param key key with which the specified value is to be associated
+     * @param value value to be associated with the specified key
+     * @return the previous value associated with {@code key}, or
+     *         {@code null} if there was no mapping for {@code key}.
+     *         (A {@code null} return can also indicate that the map
+     *         previously associated {@code null} with {@code key},
+     *         if the implementation supports {@code null} values.)
+     * @param stateFilePath The path of this state map's persistance file.
+     */
+    @Override
     public synchronized String put(final String stateFilePath, final String key, final String value) {
         Map<String, String> map = stateFileToMap.computeIfAbsent(stateFilePath, mapKey -> {
             // first check if file exists and try to load map