+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2022-12-09
+
+### Changed
+- Replaced SimpleInMemoryMapState by MapDBState; this uses memory mapped files to keep state.
+
 ## [0.1.3] - 2022-12-08
 
 ### Changed
@@ -40,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Code extracted from [RMLMapper](https://github.com/RMLio/rmlmapper-java)
 
+[0.1.4]: https://github.com/FnOio/idlab-functions-java/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/FnOio/idlab-functions-java/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/FnOio/idlab-functions-java/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/FnOio/idlab-functions-java/compare/v0.1.0...v0.1.1